Fix Terminal Tabs
