+++ CodeEdit/Features/UtilityArea/TerminalUtility/UtilityAreaTerminalTab.swift
@@ -51,8 +51,15 @@ struct UtilityAreaTerminalTab: View {
             Button("Rename...") {
                 isFocused = true
             }
-            Button("Kill Terminal") {
-                if isSelected { removeTerminals([terminal.id]) }
+
+            if selectedIDs.contains(terminal.id) && selectedIDs.count > 1 {
+                Button("Kill Terminals") {
+                    removeTerminals(selectedIDs)
+                }
+            } else {
+                Button("Kill Terminal") {
+                    removeTerminals([terminal.id])
+                }
             }
         }
     }
